create controllers

+++ mern-notes-app/backend/src/controllers/note.controller.js
@@ -1,10 +1,18 @@
 import Note from "../models/note.model.js";
+import mongoose from "mongoose";
 
 export const getNote = async (req, res) => {
 	const { id } = req.params;
 
+	if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res.status(400).json({ message: "Invalid ID!" });
+	}
+
 	try {
 		const note = await Note.findById(id);
+		if (!note) {
+			return res.status(404).json({ message: "Note not found!" });
+		}
 		res.status(200).json(note);
 	} catch (err) {
 		console.error("Error in the getNote controller!", err.message);
@@ -15,6 +23,13 @@ export const getNote = async (req, res) => {
 export const getAllNotes = async (_, res) => {
 	try {
 		const notes = await Note.find();
+
+		if (notes.length === 0) {
+			return res
+				.status(404)
+				.json({ message: "It's empty, no notes found!" });
+		}
+
 		res.status(200).json(notes);
 	} catch (err) {
 		console.error("Error in the getAll controller!", err.message);
@@ -24,6 +39,10 @@ export const getAllNotes = async (_, res) => {
 
 export const createNote = async (req, res) => {
 	const { title, description, isCompleted } = req.body;
+
+	if (!title || !description) {
+		return res.status(400).json({ message: "Please provide all fields!" });
+	}
 
 	try {
 		const newNote = await Note.create({ title, description, isCompleted });
@@ -38,6 +57,15 @@ export const updateNote = async (req, res) => {
 	const { id } = req.params;
 	const { title, description, isCompleted } = req.body;
 
+	if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res.status(400).json({ message: "Invalid ID!" });
+	}
+
+	if (!title && !description) {
+		return res
+			.status(404)
+			.json({ message: "Please provide at least one field to update!" });
+	}
 	try {
 		const updatedNote = await Note.findByIdAndUpdate(
 			id,
@@ -47,6 +75,11 @@ export const updateNote = async (req, res) => {
 				runValidators: true,
 			},
 		);
+
+		if (!updatedNote) {
+			return res.status(404).json({ message: "Note not found!" });
+		}
+
 		res.status(200).json(updatedNote);
 	} catch (err) {
 		console.error("Error in the updateNote controller!", err.message);
@@ -57,8 +90,17 @@ export const updateNote = async (req, res) => {
 export const deleteNote = async (req, res) => {
 	const { id } = req.params;
 
+	if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res.status(400).json({ message: "Invalid ID!" });
+	}
+
 	try {
-		await Note.findByIdAndDelete(id);
+		const deletedNote = await Note.findByIdAndDelete(id);
+
+		if (!deletedNote) {
+			return res.status(404).json({ message: "Note not found!" });
+		}
+
 		res.status(200).json({ message: "Note deleted successfully!" });
 	} catch (err) {
 		console.error("Error in the deleteNote controller!", err.message);
@@ -68,8 +110,14 @@ export const deleteNote = async (req, res) => {
 
 export const deleteAllNotes = async (_, res) => {
 	try {
-		await Note.deleteMany();
-		res.status(200).json({ message: "All tasks deleted successfully!" });
+		const notes = await Note.deleteMany();
+		if (notes.deletedCount === 0) {
+			return res
+				.status(404)
+				.json({ message: "It's empty, no notes deleted!" });
+		}
+
+		res.status(200).json({ message: "All notes deleted successfully!" });
 	} catch (err) {
 		console.error("Error in the deleteAllNotes controller!", err.message);
 		res.status(500).json({ message: "Internal Server Error!" });
